add BA1E. restructure everyone to use rosalind.go community function file
