+++ ba1d.go
@@ -25,30 +25,6 @@ func BA1DDescription() {
     }
 }
 
-// Given a large string (genome) and a substring (pattern),
-// find the zero-based indices where pattern occurs in genome.
-func FindOccurrences(pattern, genome string) ([]int,error) {
-    locations := []int{}
-    slots := len(genome)-len(pattern)+1
-
-    if slots<1 {
-        // pattern is longer than genome
-        return locations,nil
-    }
-
-    // Loop over each character,
-    // saving the position if it
-    // is the start of pattern
-    for i:=0; i<slots; i++ {
-        start := i
-        end := i+len(pattern)
-        if genome[start:end]==pattern {
-            locations = append(locations,i)
-        }
-    }
-    return locations,nil
-}
-
 
 // Describe the problem, and call the function
 func BA1D() {
