+++ main.go
@@ -2,7 +2,8 @@ package main
 
 func main() {
     //BA1A()
-    BA1B()
+    //BA1B()
     //BA1C()
     //BA1D()
+    BA1E()
 }
